Core/PacketIO: Updated packet structures to 9.0.5

+++ src/server/game/Server/Packets/PacketUtilities.h
@@ -103,7 +103,7 @@ namespace WorldPackets
         storage_type _storage;
     };
 
-    template<typename Underlying = uint32>
+    template<typename Underlying = int64>
     class Timestamp
     {
     public:
@@ -149,7 +149,7 @@ namespace WorldPackets
         time_t _value = time_t(0);
     };
 
-    template<typename ChronoDuration, typename Underlying = uint32>
+    template<typename ChronoDuration, typename Underlying = int64>
     class Duration
     {
     public:
